+++ packages/studio-be/src/studio/cloud/errors.ts
@@ -1,5 +1,10 @@
 import { VError } from 'verror'
 
+export class CreateBotError extends VError {
+  // @ts-ignore
+  private __brand: undefined
+}
+
 export class CDMConflictError extends VError {
   // @ts-ignore
   private __brand: undefined
